# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentBazaar API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Agent {
  id: number;
  name: string;
  description: string;
  longDescription?: string;
  category: string;
  price: number;
  rating: number;
  reviewCount: number;
  imageUrl?: string;
  tags?: string[];
  features?: string[];
  createdAt: string;
}

export interface Feedback {
  id: number;
  agentId: number;
  userName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface CreateFeedbackInput {
  agentId: number;
  userName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment: string;
}

export interface Contact {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  createdAt: string;
}

export interface CreateContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListAgentsParams = {
  category?: string;
  search?: string;
};

export type ListFeedbackParams = {
  agentId?: number;
};