# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentBazaar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Agent {
  id: number;
  name: string;
  description: string;
  longDescription?: string;
  category: string;
  price: number;
  rating: number;
  reviewCount: number;
  imageUrl?: string;
  tags?: string[];
  features?: string[];
  createdAt: string;
}