# lib/api-zod/src/generated/types/createContactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentBazaar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}