# lib/api-zod/src/generated/types/createFeedbackInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentBazaar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFeedbackInput {
  agentId: number;
  userName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment: string;
}