# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgentBazaar API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  agentId: number;
  userName: string;
  rating: number;
  comment: string;
  createdAt: string;
}